@@ -5,29 +5,25 @@ import 'package:flutter_github_scripts/github_datatypes.dart';
 import 'package:flutter_github_scripts/github_queries.dart';
 
 class Options {
-  final _parser = ArgParser(allowTrailingOptions: false);
-  ArgResults _results;
-  bool get showClosed => _results['closed'];
-  bool get showMerged => _results['merged'];
-  DateTime get from =>
-      _results.rest != null ? DateTime.parse(_results.rest[0]) : null;
-  DateTime get to =>
-      _results.rest != null ? DateTime.parse(_results.rest[1]) : null;
-  bool get labels => _results['labels'];
-  bool get authors => _results['authors'];
-  bool get assignees => _results['assignees'];
-  bool get reviewers => _results['reviewers'];
-  bool get prs => _results['prs'];
-  bool get issues => _results['issues'];
-  bool get alphabetize => _results['alphabetize'];
-  bool get customers => _results['customers-only'];
-  bool get ranking => _results['ranking'];
-  bool get skipUninteresting => _results['skip-uninteresting-labels'];
-  int get exitCode => _results == null
-      ? -1
-      : _results['help']
-          ? 0
-          : null;
+  final ArgParser _parser = ArgParser(allowTrailingOptions: false);
+  late ArgResults _results;
+  bool? get showClosed => _results['closed'];
+  bool? get showMerged => _results['merged'];
+  DateTime? get from =>
+      _results.rest.isNotEmpty ? DateTime.parse(_results.rest[0]) : null;
+  DateTime? get to =>
+      _results.rest.length >= 2 ? DateTime.parse(_results.rest[1]) : null;
+  bool? get labels => _results['labels'];
+  bool? get authors => _results['authors'];
+  bool? get assignees => _results['assignees'];
+  bool? get reviewers => _results['reviewers'];
+  bool? get prs => _results['prs'];
+  bool? get issues => _results['issues'];
+  bool? get alphabetize => _results['alphabetize'];
+  bool? get customers => _results['customers-only'];
+  bool? get ranking => _results['ranking'];
+  bool? get skipUninteresting => _results['skip-uninteresting-labels'];
+  int? get exitCode => _results['help'] ? 0 : null;
 
   Options(List<String> args) {
     _parser
@@ -125,24 +121,25 @@ class Options {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
-  var keys = Set<String>();
+  if (opts.exitCode != null) exit(opts.exitCode!);
+  var keys = Set<String?>();
 
-  final repos = opts.prs ? ['flutter', 'engine', 'plugins'] : ['flutter'];
+  final repos = opts.prs! ? ['flutter', 'engine', 'plugins'] : ['flutter'];
   final labelsToSkip = ['cla: yes', 'waiting for tree to go green'];
 
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
-  GitHubIssueType type;
-  if (opts.issues) type = GitHubIssueType.issue;
-  if (opts.prs) type = GitHubIssueType.pullRequest;
+  GitHubIssueType? type;
+  if (opts.issues!) type = GitHubIssueType.issue;
+  if (opts.prs!) type = GitHubIssueType.pullRequest;
 
   var state = GitHubIssueState.open;
-  DateRange when = null;
+  DateRange? when = null;
   var rangeType = GitHubDateQueryType.none;
-  if (opts.showClosed || opts.showMerged) {
-    state = opts.showClosed ? GitHubIssueState.closed : GitHubIssueState.merged;
+  if (opts.showClosed! || opts.showMerged!) {
+    state =
+        opts.showClosed! ? GitHubIssueState.closed : GitHubIssueState.merged;
     when = DateRange(DateRangeType.range, start: opts.from, end: opts.to);
     rangeType = GitHubDateQueryType.closed;
   }
@@ -156,35 +153,35 @@ void main(List<String> args) async {
         dateQuery: rangeType,
         dateRange: when);
 
-    Cluster clusters;
-    if (opts.labels) clusters = Cluster.byLabel(items);
-    if (opts.authors) clusters = Cluster.byAuthor(items);
-    if (opts.assignees) clusters = Cluster.byAssignees(items);
-    if (opts.reviewers) clusters = Cluster.byReviewers(items);
+    late Cluster clusters;
+    if (opts.labels!) clusters = Cluster.byLabel(items);
+    if (opts.authors!) clusters = Cluster.byAuthor(items);
+    if (opts.assignees!) clusters = Cluster.byAssignees(items);
+    if (opts.reviewers!) clusters = Cluster.byReviewers(items);
 
     for (var key in clusters.clusters.keys) {
       keys.add(key);
     }
 
     var what = '';
-    if (opts.authors) what = 'authors';
-    if (opts.labels) what = 'labels';
-    if (opts.assignees) what = 'owners';
+    if (opts.authors!) what = 'authors';
+    if (opts.labels!) what = 'labels';
+    if (opts.assignees!) what = 'owners';
 
     var reportType = 'Open';
-    if (opts.showMerged) reportType = 'Merged';
-    if (opts.showClosed) reportType = 'Closed';
+    if (opts.showMerged!) reportType = 'Merged';
+    if (opts.showClosed!) reportType = 'Closed';
     print('### ${reportType} ' +
-        (opts.issues ? 'issues' : 'PRs') +
+        (opts.issues! ? 'issues' : 'PRs') +
         ' by ${what}' +
         ' for `flutter/${repo}` ' +
-        (opts.showClosed
-            ? 'from ${opts.from.toIso8601String()} to ${opts.to.toIso8601String()}'
+        (opts.showClosed!
+            ? 'from ${opts.from!.toIso8601String()} to ${opts.to!.toIso8601String()}'
             : '') +
         '\n\n');
 
-    if (opts.customers) {
-      Set<String> toRemove = Set<String>();
+    if (opts.customers!) {
+      Set<String?> toRemove = Set<String?>();
       for (var label in clusters.clusters.keys) {
         if (label.indexOf('customer: ') != 0) toRemove.add(label);
       }
@@ -193,8 +190,8 @@ void main(List<String> args) async {
       }
     }
 
-    if (opts.labels && opts.skipUninteresting) {
-      Set<String> toRemove = Set<String>();
+    if (opts.labels! && opts.skipUninteresting!) {
+      Set<String?> toRemove = Set<String?>();
       for (var label in clusters.clusters.keys) {
         if (labelsToSkip.contains(label)) toRemove.add(label);
       }
@@ -204,38 +201,39 @@ void main(List<String> args) async {
     }
 
     print(clusters.toMarkdown(
-        sortType: (opts.alphabetize
+        sortType: (opts.alphabetize!
             ? ClusterReportSort.byKey
             : ClusterReportSort.byCount),
         skipEmpty: true,
         showStatistics: false));
 
-    if (opts.authors) {
+    if (opts.authors!) {
       print('${clusters.clusters.keys.length} unique ' +
-          (opts.labels ? 'labels.' : 'users') +
+          (opts.labels! ? 'labels.' : 'users') +
           ' across this repository.\n\n');
     }
 
-    if (opts.ranking) {
+    if (opts.ranking!) {
       print('### Customer ' +
-          (opts.issues ? 'issues' : 'PRs') +
+          (opts.issues! ? 'issues' : 'PRs') +
           ' rank-ordered by label');
       for (var customer in clusters.clusters.keys) {
-        var labelCountsByLabel = Map<String, int>();
+        var labelCountsByLabel = Map<String?, int>();
         for (var item in clusters.clusters[customer]) {
           for (var labelItem in item.labels.labels) {
             var label = labelItem as Label;
             if (!labelCountsByLabel.containsKey(label.label)) {
               labelCountsByLabel[label.label] = 1;
             } else {
-              labelCountsByLabel[label.label]++;
+              labelCountsByLabel[label.label] =
+                  labelCountsByLabel[label.label]! + 1;
             }
           }
         }
         var rankedLabelList = labelCountsByLabel.keys.toList();
         // REVERSE sort, not incremental sort
         rankedLabelList.sort(
-            (a, b) => labelCountsByLabel[b].compareTo(labelCountsByLabel[a]));
+            (a, b) => labelCountsByLabel[b]!.compareTo(labelCountsByLabel[a]!));
 
         print('#### ${customer}\n\n');
         for (var labelName in rankedLabelList) {
@@ -245,9 +243,9 @@ void main(List<String> args) async {
     }
   }
 
-  if (opts.authors) {
+  if (opts.authors!) {
     print('A total of ${keys.length} unique ' +
-        (opts.labels ? 'labels' : 'users') +
+        (opts.labels! ? 'labels' : 'users') +
         ' across all repositories.\n\n');
   }
 }
