@@ -6,15 +6,15 @@ import 'package:flutter_github_scripts/github_queries.dart';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  ArgResults _results;
-  String get login => _results.rest[0];
-  bool get alwaysIncludeTeam => _results['always-include-team'];
-  int get exitCode => _results == null
+  ArgResults? _results;
+  String get login => _results!.rest[0];
+  bool? get alwaysIncludeTeam => _results!['always-include-team'];
+  int? get exitCode => _results == null
       ? -1
-      : _results['help']
+      : _results!['help']
           ? 0
           : null;
-  bool get tsv => _results['tsv'];
+  bool? get tsv => _results!['tsv'];
   Options(List<String> args) {
     _parser
       ..addFlag('help',
@@ -26,8 +26,8 @@ class Options {
           help: 'include team name in each row.');
     try {
       _results = _parser.parse(args);
-      if (_results['help']) _printUsage();
-      if (_results.rest.length != 1) throw ('invalid organization!');
+      if (_results!['help']) _printUsage();
+      if (_results!.rest.length != 1) throw ('invalid organization!');
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
@@ -43,11 +43,11 @@ class Options {
 }
 
 class MemberInfo {
-  String _login;
+  String? _login;
   get login => _login;
-  DateTime firstContributed;
+  DateTime? firstContributed;
   dynamic firstContribution;
-  DateTime lastContributed;
+  DateTime? lastContributed;
   dynamic lastContribution;
 
   MemberInfo(this._login);
@@ -55,8 +55,8 @@ class MemberInfo {
 
 enum When { first, last }
 
-findWhen(dynamic item, String login, When w) {
-  DateTime result;
+findWhen(dynamic item, String? login, When w) {
+  DateTime? result;
   switch (w) {
     case When.first:
       if (item.author?.login == login) {
@@ -96,16 +96,17 @@ findWhen(dynamic item, String login, When w) {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
   // Enumerate all of the teams and get all of the members of all of the teams.
   var org = await github.organization(opts.login);
-  var allMembers = Map<String, MemberInfo>();
-  var membersByTeam = SplayTreeMap<String, List<String>>();
+  var allMembers = Map<String?, MemberInfo>();
+  SplayTreeMap<String?, List<String?>> membersByTeam =
+      SplayTreeMap<String?, List<String>>();
   await for (var team in org.teamsStream) {
-    var membersThisTeam = <String>[];
+    var membersThisTeam = <String?>[];
     await for (var member in team.membersStream) {
       membersThisTeam.add(member.login);
       allMembers[member.login] = MemberInfo(member.login);
@@ -137,20 +138,20 @@ void main(List<String> args) async {
     }
 
     if (earliestAuthored == null && earliestCommented == null) {
-      member.firstContributed = null;
+      member!.firstContributed = null;
       member.firstContribution = null;
     } else if (earliestAuthored == null && earliestCommented != null) {
-      member.firstContributed = findWhen(earliestCommented, login, When.first);
+      member!.firstContributed = findWhen(earliestCommented, login, When.first);
       member.firstContribution = earliestCommented;
     } else if (earliestAuthored != null && earliestCommented == null) {
-      member.firstContributed = findWhen(earliestAuthored, login, When.first);
+      member!.firstContributed = findWhen(earliestAuthored, login, When.first);
       member.firstContribution = earliestAuthored;
     } else if (findWhen(earliestCommented, login, When.first)
         .isBefore(findWhen(earliestAuthored, login, When.first))) {
-      member.firstContributed = findWhen(earliestCommented, login, When.first);
+      member!.firstContributed = findWhen(earliestCommented, login, When.first);
       member.firstContribution = earliestCommented;
     } else {
-      member.firstContributed = findWhen(earliestAuthored, login, When.first);
+      member!.firstContributed = findWhen(earliestAuthored, login, When.first);
       member.firstContribution = earliestAuthored;
     }
 
@@ -195,10 +196,10 @@ void main(List<String> args) async {
   print(
       'Team\tGithub login\tFirst contributed\tEarliest contribution\tLast contributed\tLatest contribution');
   for (var team in membersByTeam.keys) {
-    if (!opts.alwaysIncludeTeam) print('${team}');
-    for (var member in membersByTeam[team]) {
-      var row = opts.alwaysIncludeTeam ? '${team}\t${member}' : '\t${member}';
-      var contributor = allMembers[member];
+    if (!opts.alwaysIncludeTeam!) print('${team}');
+    for (var member in membersByTeam[team]!) {
+      var row = opts.alwaysIncludeTeam! ? '${team}\t${member}' : '\t${member}';
+      var contributor = allMembers[member]!;
       if (contributor.firstContribution == null) {
         row += '\t\t';
       } else {
