@@ -6,15 +6,15 @@ import 'package:flutter_github_scripts/github_queries.dart';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  ArgResults _results;
-  DateTime get from => DateTime.parse(_results['from']);
-  DateTime get to => DateTime.parse(_results['to']);
-  bool get showQueries => _results['queries'];
+  ArgResults? _results;
+  DateTime get from => DateTime.parse(_results!['from']);
+  DateTime get to => DateTime.parse(_results!['to']);
+  bool? get showQueries => _results!['queries'];
   int get deltaDays =>
-      int.parse(_results['delta'] == null ? '7' : _results['delta']);
-  int get exitCode => _results == null
+      int.parse(_results!['delta'] == null ? '7' : _results!['delta']);
+  int? get exitCode => _results == null
       ? -1
-      : _results['help']
+      : _results!['help']
           ? 0
           : null;
   Options(List<String> args) {
@@ -36,7 +36,7 @@ class Options {
 
     try {
       _results = _parser.parse(args);
-      if (_results['help']) _printUsage();
+      if (_results!['help']) _printUsage();
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
@@ -52,9 +52,11 @@ class Options {
 
 class MeanComputer {
   // Running mean of all invocations
-  double _totalSumSeconds;
+  late double _totalSumSeconds;
   get totalSum => _totalSumSeconds;
-  double _totalCount;
+
+  late double _totalCount;
+
   get totalCount => _totalCount;
   get meanDuration => _totalCount == 0
       ? Duration(seconds: 0)
@@ -65,17 +67,18 @@ class MeanComputer {
     _totalCount = 0.0;
   }
 
-  Duration computeMean(
-    List<dynamic> issues,
-  ) {
+  Duration computeMean(List<dynamic> issues) {
+    // TODO: note that onlyCustomers is only ever false, and creates dead code
+    // below.
     bool onlyCustomers = false;
     double sum = 0.0;
-    var count = issues == null ? 0 : issues.length;
+    var count = issues.length;
     if (count == 0) return Duration(seconds: 0);
     bool hasCustomer = false;
     for (var item in issues) {
       var issue = item as Issue;
       if (issue.closedAt == null) continue;
+      // ignore: dead_code
       if (onlyCustomers) {
         for (var label in issue.labels.labels) {
           if (label.label.contains('customer:')) {
@@ -84,6 +87,7 @@ class MeanComputer {
           }
         }
       }
+      // ignore: dead_code
       if (!onlyCustomers || (onlyCustomers && hasCustomer)) {
         var delta = issue.closedAt.difference(issue.createdAt);
         sum += delta.inSeconds;
@@ -104,7 +108,7 @@ class MeanComputer {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
@@ -126,7 +130,7 @@ void main(List<String> args) async {
 
     var closedQuery =
         'repo:FirebaseExtended/flutterfire is:issue sort:updated-desc closed:${fromStamp}..${toStamp}';
-    if (opts.showQueries) {
+    if (opts.showQueries!) {
       print(openQuery);
       print(closedQuery);
     }
@@ -134,8 +138,8 @@ void main(List<String> args) async {
     var openIssues = github.searchIssuePRs(openQuery);
     var closedIssues = github.searchIssuePRs(closedQuery);
 
-    List<Issue> openedThisPeriod = [];
-    List<Issue> closedThisPeriod = [];
+    List<Issue?> openedThisPeriod = [];
+    List<Issue?> closedThisPeriod = [];
 
     await for (var issue in openIssues) {
       if (issue.createdAt.compareTo(opts.from) >= 0 &&
