@@ -7,17 +7,17 @@ import 'package:flutter_github_scripts/github_queries.dart';
 
 class Options {
   final _parser = ArgParser(allowTrailingOptions: false);
-  ArgResults _results;
-  DateTime get from => DateTime.parse(_results['from']);
-  DateTime get to => DateTime.parse(_results['to']);
-  bool get summarize => _results['summarize'];
-  bool get showQueries => _results['queries'];
-  bool get onlyCustomers => _results['customers'];
+  ArgResults? _results;
+  DateTime get from => DateTime.parse(_results!['from']);
+  DateTime get to => DateTime.parse(_results!['to']);
+  bool? get summarize => _results!['summarize'];
+  bool? get showQueries => _results!['queries'];
+  bool? get onlyCustomers => _results!['customers'];
   int get deltaDays =>
-      int.parse(_results['delta'] == null ? '7' : _results['delta']);
-  int get exitCode => _results == null
+      int.parse(_results!['delta'] == null ? '7' : _results!['delta']);
+  int? get exitCode => _results == null
       ? -1
-      : _results['help']
+      : _results!['help']
           ? 0
           : null;
   Options(List<String> args) {
@@ -47,7 +47,7 @@ class Options {
 
     try {
       _results = _parser.parse(args);
-      if (_results['help']) _printUsage();
+      if (_results!['help']) _printUsage();
     } on ArgParserException catch (e) {
       print(e.message);
       _printUsage();
@@ -61,7 +61,10 @@ class Options {
   }
 }
 
-int countWithOrWithoutCustomers(List<dynamic> issues, {bool onlyCustomers}) {
+int countWithOrWithoutCustomers(
+  List<dynamic> issues, {
+  required bool onlyCustomers,
+}) {
   var count = issues.length;
   if (onlyCustomers) {
     bool hasCustomer = false;
@@ -83,9 +86,9 @@ int countWithOrWithoutCustomers(List<dynamic> issues, {bool onlyCustomers}) {
 
 class MeanComputer {
   // Running mean of all invocations
-  double _totalSumSeconds;
+  late double _totalSumSeconds;
   get totalSum => _totalSumSeconds;
-  double _totalCount;
+  late double _totalCount;
   get totalCount => _totalCount;
   get meanDuration => _totalCount == 0
       ? Duration(seconds: 0)
@@ -96,16 +99,16 @@ class MeanComputer {
     _totalCount = 0.0;
   }
 
-  Duration meanDurationWithOrWithoutCustomers(List<dynamic> issues,
-      {bool onlyCustomers}) {
+  Duration meanDurationWithOrWithoutCustomers(List<dynamic>? issues,
+      {bool? onlyCustomers}) {
     double sum = 0.0;
     var count = issues == null ? 0 : issues.length;
     if (count == 0) return Duration(seconds: 0);
     bool hasCustomer = false;
-    for (var item in issues) {
+    for (var item in issues!) {
       var issue = item as Issue;
       if (issue.closedAt == null) continue;
-      if (onlyCustomers) {
+      if (onlyCustomers!) {
         for (var label in issue.labels.labels) {
           if (label.label.contains('customer:')) {
             hasCustomer = true;
@@ -133,17 +136,17 @@ class MeanComputer {
 
 void main(List<String> args) async {
   final opts = Options(args);
-  if (opts.exitCode != null) exit(opts.exitCode);
+  if (opts.exitCode != null) exit(opts.exitCode!);
   final token = Platform.environment['GITHUB_TOKEN'];
   final github = GitHub(token);
 
   DateTime current = opts.from, last = opts.to;
 
-  if (opts.summarize) {
+  if (opts.summarize!) {
     print(
         'This shows the number of new, open, and closed `P0`, `P1`, and `P2` issues over the period from');
     print('${opts.from} to ${opts.to}.');
-    if (opts.onlyCustomers) {
+    if (opts.onlyCustomers!) {
       print('Only issues with at least one `customer` label are presented.');
     }
     print('Period ending\tCreated P0s\tCreated P1s\tCreated P2s\t' +
@@ -162,7 +165,7 @@ void main(List<String> args) async {
 
     var closedQuery =
         'org:flutter is:issue sort:updated-desc closed:${fromStamp}..${toStamp}';
-    if (opts.showQueries) {
+    if (opts.showQueries!) {
       print(openQuery);
       print(closedQuery);
     }
@@ -170,8 +173,8 @@ void main(List<String> args) async {
     var openIssues = github.searchIssuePRs(openQuery);
     var closedIssues = github.searchIssuePRs(closedQuery);
 
-    List<Issue> openedThisPeriod = [];
-    List<Issue> closedThisPeriod = [];
+    List<Issue?> openedThisPeriod = [];
+    List<Issue?> closedThisPeriod = [];
 
     await for (var issue in openIssues) {
       if (issue.createdAt.compareTo(opts.from) >= 0 &&
@@ -189,7 +192,7 @@ void main(List<String> args) async {
     final interestingPriorities = ['P0', 'P1', 'P2'];
     interestingPriorities.sort();
 
-    if (opts.summarize) {
+    if (opts.summarize!) {
       var openCount = HashMap<String, int>();
       var closeCount = HashMap<String, int>();
       var meanComputerUntilClosed = MeanComputer();
@@ -204,11 +207,11 @@ void main(List<String> args) async {
         openCount[p] = highPrioritizedIssuesOpened == null
             ? 0
             : countWithOrWithoutCustomers(highPrioritizedIssuesOpened,
-                onlyCustomers: opts.onlyCustomers);
+                onlyCustomers: opts.onlyCustomers!);
         closeCount[p] = highPrioritizedIssuesClosed == null
             ? 0
             : countWithOrWithoutCustomers(highPrioritizedIssuesClosed,
-                onlyCustomers: opts.onlyCustomers);
+                onlyCustomers: opts.onlyCustomers!);
         meanUntilClosed[p] = meanComputerUntilClosed
             .meanDurationWithOrWithoutCustomers(highPrioritizedIssuesClosed,
                 onlyCustomers: opts.onlyCustomers);
@@ -222,9 +225,9 @@ void main(List<String> args) async {
       interestingPriorities.forEach((p) => row = '${row}\t${openCount[p]}');
       interestingPriorities.forEach((p) => row = '${row}\t${closeCount[p]}');
       interestingPriorities
-          .forEach((p) => row = '${row}\t${meanUntilClosed[p].inHours}');
+          .forEach((p) => row = '${row}\t${meanUntilClosed[p]!.inHours}');
       interestingPriorities
-          .forEach((p) => row = '${row}\t${meanOpenClosed[p].inHours}');
+          .forEach((p) => row = '${row}\t${meanOpenClosed[p]!.inHours}');
       row = '${row}\t${meanComputerUntilClosed.meanDuration.inHours}';
       row = '${row}\t${meanComputerOpenClosed.meanDuration.inHours}';
       print(row);
@@ -232,7 +235,7 @@ void main(List<String> args) async {
       print(
           'This shows the number of new, open, and closed `P0`, `P1`, and `P2` issues over the period from');
       print('${fromStamp} to ${toStamp}.\n\n');
-      if (opts.onlyCustomers) {
+      if (opts.onlyCustomers!) {
         print('Only issues with at least one `customer` label are presented.');
       }
 
@@ -247,11 +250,11 @@ void main(List<String> args) async {
         var openCount = highPrioritizedIssuesOpened == null
             ? 0
             : countWithOrWithoutCustomers(highPrioritizedIssuesOpened,
-                onlyCustomers: opts.onlyCustomers);
+                onlyCustomers: opts.onlyCustomers!);
         var closedCount = highPrioritizedIssuesClosed == null
             ? 0
             : countWithOrWithoutCustomers(highPrioritizedIssuesClosed,
-                onlyCustomers: opts.onlyCustomers);
+                onlyCustomers: opts.onlyCustomers!);
         var totalRow = openCount + closedCount;
         totalOpen += openCount;
         totalClosed += closedCount;
